created question.py for dataclasses
